Refactor mqtt inbound handler.
